[Oauth] WIP Github register almost done (#6)

+++ src/domain/user/useCases/createUser/createUser.service.ts
@@ -4,6 +4,7 @@ import { MailerService } from '@nestjs-modules/mailer';
 import * as bcrypt from 'bcrypt';
 import * as crypto from 'crypto'
 import { RedisHandlerService } from '../../services/auth/redis/redis-handler.service';
+import { CreateUserDto } from '../../dtos/createUser.dto';
 
 @Injectable()
 export class CreateUserService {
@@ -13,42 +14,62 @@ export class CreateUserService {
         private readonly mailerService: MailerService
     ) { }
 
-    async create(body) {
+    async create(body, oauth = false) {
 
         let user;
 
         try {
-            const hashedPassword = await bcrypt.hash(body.password, 10);
-            body.password = hashedPassword
+
+            if (!oauth) {
+                const hashedPassword = await bcrypt.hash(body.password, 10);
+                body.password = hashedPassword
+            }
 
             user = await this.usersRepository.createUser(body);
 
             if (user) {
-                const confirmEmailToken = crypto.randomBytes(64).toString('hex');
+                if (!body.confirmed) {
+                    const confirmEmailToken = crypto.randomBytes(64).toString('hex');
 
-                const key = process.env.CONFIRM_EMAIL_PREFIX + confirmEmailToken
+                    const key = process.env.CONFIRM_EMAIL_PREFIX + confirmEmailToken
 
-                await this.redisHandlerService.client.set(
-                    key,
-                    user.id,
-                    'ex',
-                    process.env.CONFIRM_EMAIL_TOKEN_EXPIRES_IN
-                )
+                    await this.redisHandlerService.client.set(
+                        key,
+                        user.id,
+                        'ex',
+                        process.env.CONFIRM_EMAIL_TOKEN_EXPIRES_IN
+                    )
 
-                try {
-                    await this.mailerService
-                        .sendMail({
-                            to: user.email,
-                            from: 'ton@pote.dev', // sender address
-                            subject: 'Inscription sur pote.dev ✔', // Subject line
-                            text: `Bienvenue ${user.firstname} sur pote.dev, please confirm email by clicking <a href="${process.env.APP_BASE_URL}/confirm_email/${confirmEmailToken}">confirm email</a> `, // plaintext body
-                            html: `<b>Bienvenue ${user.firstname} sur pote.dev, please confirm email by clicking <a href="${process.env.APP_BASE_URL}/confirm_email/${confirmEmailToken}">confirm email</a></b>`, // HTML body content
-                        })
-                }
-                catch (err) {
-                    //If email isn't sent
-                    await this.redisHandlerService.client.del(key)
-                    throw new BadRequestException("Wrong email format");
+                    try {
+                        await this.mailerService
+                            .sendMail({
+                                to: user.email,
+                                from: 'ton@pote.dev', // sender address
+                                subject: 'Inscription sur pote.dev ✔', // Subject line
+                                text: `Bienvenue ${user.username} sur pote.dev, please confirm email by clicking <a href="${process.env.APP_BASE_URL}/confirm_email/${confirmEmailToken}">confirm email</a> `, // plaintext body
+                                html: `<b>Bienvenue ${user.username} sur pote.dev, please confirm email by clicking <a href="${process.env.APP_BASE_URL}/confirm_email/${confirmEmailToken}">confirm email</a></b>`, // HTML body content
+                            })
+                    }
+                    catch (err) {
+                        //If email isn't sent
+                        await this.redisHandlerService.client.del(key)
+                        throw new BadRequestException("Wrong email format");
+                    }
+                } else {
+                    try {
+                        await this.mailerService
+                            .sendMail({
+                                to: user.email,
+                                from: 'ton@pote.dev', // sender address
+                                subject: 'Inscription sur pote.dev ✔', // Subject line
+                                text: `Bienvenue ${user.username} sur pote.dev`, // plaintext body
+                                html: `<b>Bienvenue ${user.username} sur pote.dev`, // HTML body content
+                            })
+                    }
+                    catch (err) {
+                        //If email isn't sent
+                        throw new BadRequestException("Wrong email format");
+                    }
                 }
             }
         }
